separation of concern done

+++ frontend/src/services/authService.js
@@ -1,23 +1,24 @@
-import API from "./api";
+import API, { publicAPI } from "./api";
 
 export const registerUser = (data) => {
-  return API.post("auth/register/", data);
+  return publicAPI.post("auth/register/", data);
 };
 
 export const loginUser = (data) => {
-  return API.post("auth/login/", data);
+  return publicAPI.post("auth/login/", data);
 };
 
 export const getProfile = () => {
-  return API.get("profile/");
+  return API.get("profile/", { requiresAuth: true });
 };
 
 export const getUsers = () => {
-  return API.get("users/");
+  return API.get("users/", { requiresAuth: true });
 };
 
 export const updateProfile = (data) => {
   return API.patch("profile/", data, {
+    requiresAuth: true,
     headers: {
       "Content-Type": "multipart/form-data",
     },
@@ -26,6 +27,7 @@ export const updateProfile = (data) => {
 
 export const updateUser = (id, data) => {
   return API.patch(`users/${id}/`, data, {
+    requiresAuth: true,
     headers: {
       "Content-Type": "multipart/form-data",
     },
@@ -33,21 +35,22 @@ export const updateUser = (id, data) => {
 };
 
 export const deleteAccount = () => {
-  return API.delete("profile/");
+  return API.delete("profile/", { requiresAuth: true });
 };
 
 export const deleteUser = (id) => {
-  return API.delete(`users/${id}/`);
+  return API.delete(`users/${id}/`, { requiresAuth: true });
 };
 
+// ❌ NO AUTH
 export const sendOTP = (data) => {
-  return API.post("auth/send-otp/", data);
+  return publicAPI.post("auth/send-otp/", data);
 };
 
 export const verifyOTP = (data) => {
-  return API.post("auth/verify-otp/", data);
+  return publicAPI.post("auth/verify-otp/", data);
 };
 
 export const resetPassword = (data) => {
-  return API.post("auth/reset-password/", data);
+  return publicAPI.post("auth/reset-password/", data);
 };
